+++ examples/resources/confluencedc_global_permission/import.sh
@@ -1,0 +1,1 @@
+terraform import confluencedc_global_permission.example <group_name>/<operation_key>/<operation_target>